poetry/pyproject.toml
[tool.poetry]
name = "poetry"
version = "0.1.0"
description = ""
authors = ["luciheit@gmail.com <luciheit@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.1.6"
djangorestframework = "^3.12.2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
